Extract git-blame AttributionSource adapter into production module
Moves parseBlamePorcelain / buildBlameRuns / makeGitBlameSource out
of the test file into attribution-gitblame.ts so the adapter can be
wired into production. Synthetic unit tests added alongside the
existing end-to-end.
